Adds tests for argument parsing

+++ tests/conftest.py
@@ -1,6 +1,9 @@
 """Set of fixtures to be used by the tests."""
 # pylint: disable=redefined-outer-name
+import json
 import os
+import random
+from collections import OrderedDict
 from shutil import rmtree
 
 import mock
@@ -19,41 +22,60 @@ def link_mapping():
         'tests/assets/src/zshrc': 'tests/assets/target/.zshrc',
     }
 
-    return _transform_paths(mappings)
+    yield _transform_paths(mappings)
+
+    _cleanup_assets()
 
 
 @pytest.fixture
-def copy_mapping():
+def copy_file_mapping():
     """Returns the mappings needed to copy files with dotty."""
     mappings = {
-        'tests/assets/src/init.vim': 'tests/assets/target/init.vim',
-        'tests/assets/src/bin': 'tests/assets/target/bin',
+        'tests/assets/src/init.vim': 'tests/assets/target/.init.vim',
+        'tests/assets/src/offlineimaprc': 'tests/assets/target/.offlineimaprc',
     }
 
-    return _transform_paths(mappings)
+    yield _transform_paths(mappings)
+
+    _cleanup_assets()
+
+
+@pytest.fixture
+def copy_folder_mapping():
+    """Returns the mappings needed to copy folders with dotty."""
+    mappings = {
+        'tests/assets/src/bin': 'tests/assets/target/.bin',
+        'tests/assets/src/copy_dir': 'tests/assets/target/.copy_dir',
+    }
+
+    yield _transform_paths(mappings)
+
+    _cleanup_assets()
 
 
 @pytest.fixture
 def directory_list():
     """Returns a list of directories to create."""
     yield _transform_paths([
-        'tests/assets/target/dir1',
-        'tests/assets/target/dir2',
-        'tests/assets/target/dir3',
+        'tests/assets/target/.dir1',
+        'tests/assets/target/.dir2',
+        'tests/assets/target/.dir3',
     ])
 
-    rmtree(ASSETS_DIR)
+    _cleanup_assets()
 
 
 @pytest.fixture
 def git_repo_mapping():
     """Returns a mapping of Git repos to clone."""
-    return _transform_paths({
+    yield _transform_paths({
         'https://github.com/vibhavp/dotty.git': 'tests/assets/target/dotty',
         'https://github.com/robbyrussell/oh-my-zsh.git': (
             'tests/assets/target/omz'
         ),
     })
+
+    _cleanup_assets()
 
 
 @pytest.fixture
@@ -102,22 +124,8 @@ def package_list(request):
     patch.stop()
 
 
-def _transform_paths(mappings):
-    """Transforms the values in a provided mapping to a full path relative to
-    the current directory.
-    """
-    if isinstance(mappings, dict):
-        for key, value in mappings.items():
-            mappings[key] = os.path.abspath(value)
-    else:
-        for idx, value in enumerate(mappings):
-            mappings[idx] = os.path.abspath(value)
-
-    return mappings
-
-
 @pytest.fixture
-def assets(link_mapping, copy_mapping):
+def assets(link_mapping, copy_file_mapping):
     """Fixture that will create the tests assets and delete them when finished.
     """
     for folder in ('src', 'target'):
@@ -125,11 +133,65 @@ def assets(link_mapping, copy_mapping):
         if not os.path.exists(folder):
             os.makedirs(folder)
 
-    for mapping in (link_mapping, copy_mapping):
+    for mapping in (link_mapping, copy_file_mapping):
         for test_file in mapping.keys():
             with open(test_file, 'w') as filehandler:
                 filehandler.write('a')
 
     yield {'pytest': 'is silly about yield'}
 
-    rmtree(ASSETS_DIR)
+    _cleanup_assets()
+
+
+@pytest.fixture
+def full_mapping(link_mapping, copy_file_mapping, directory_list,
+                 git_repo_mapping, command_list, package_list):
+    """Creates full dotty mapping in an OrderedDict."""
+    return OrderedDict((
+        ('directories', directory_list),
+        ('link', link_mapping),
+        ('copy', copy_file_mapping),
+        ('commands', command_list),
+        ('git_repos', git_repo_mapping),
+        (package_list[1], package_list[0])
+    ))
+
+
+@pytest.fixture
+def dotty_json_file(full_mapping):
+    """Creates a dotty test JSON file in the asserts dir."""
+    # Create the folder
+    folder_path = _transform_paths(os.path.join(ASSETS_DIR, 'src'))
+    os.makedirs(folder_path)
+
+    # Write the JSON file
+    path = os.path.join(folder_path, 'dotty.json')
+
+    with open(path, 'w') as json_fp:
+        json.dump(full_mapping, json_fp)
+
+    yield path
+
+    _cleanup_assets()
+
+
+def _transform_paths(mappings):
+    """Transforms the values in a provided mapping to a full path relative to
+    the current directory.
+    """
+    if isinstance(mappings, dict):
+        for key, value in mappings.items():
+            mappings[key] = os.path.abspath(value)
+    elif isinstance(mappings, (list, tuple)):
+        for idx, value in enumerate(mappings):
+            mappings[idx] = os.path.abspath(value)
+    else:
+        return os.path.abspath(mappings)
+
+    return mappings
+
+
+def _cleanup_assets():
+    """Deletes the assets created by the tests."""
+    if os.path.exists(ASSETS_DIR):
+        rmtree(ASSETS_DIR)
